File-based persistence.

+++ src/main/resources/db/migration/V1_0__init.sql
@@ -1,5 +1,5 @@
 create table PxeInstall (
-    id int not null primary key auto_increment,
+    id bigint not null primary key auto_increment,
     uuid varchar(50) not null,
     macAddress varchar(20) not null
 );
